--- a/lib/diet.ml
+++ b/lib/diet.ml
@@ -90,29 +90,27 @@ module Make(Elt: ELT) = struct
     | Node: node -> t
   and node = { x: elt; y: elt; l: t; r: t; h: int; cardinal: elt }
 
-  type enum = End | More of interval * t * enum
-
   let rec cons_enum t enum =
     match t with
     | Empty -> enum
-    | Node {x; y; l; r; _} -> cons_enum l (More ((x, y), r, enum))
+    | Node ({l; _} as node) -> cons_enum l (node::enum)
 
-  let compare_with_invariant (x, y) (x', y') =
+  let compare_with_invariant {x; y; _} {x = x'; y = y'; _} =
     if eq x x' && eq y y' then 0
     else if y < x' then -1
     else 1
 
   let rec compare_aux enum enum' =
     match enum, enum' with
-    | End, End -> 0
-    | End, _ -> -1
-    | _, End -> 1
-    | More (interval, r, enum), More (interval', r', enum') ->
-      (match compare_with_invariant interval interval' with
-       | 0 -> compare_aux (cons_enum r enum) (cons_enum r' enum')
+    | [], [] -> 0
+    | [], _ -> -1
+    | _, [] -> 1
+    | node::enum, node'::enum' ->
+      (match compare_with_invariant node node' with
+       | 0 -> compare_aux (cons_enum node.r enum) (cons_enum node'.r enum')
        | c -> c)
 
-  let compare t t' = compare_aux (cons_enum t End) (cons_enum t' End)
+  let compare t t' = compare_aux (cons_enum t []) (cons_enum t' [])
 
   let equal t t' = compare t t' = 0