Fixes #7 

This PR adds `compare` and `equal` functions.

Implementation is the same as stdlib's `Set.compare`. Due to the stronger invariants in DIET, I think it might be possible to optimize it further by traversing both trees and stopping if two incompatible intervals (ie which intersect or whose bounds are adjacent) are met but I felt like this was a good enough first version.

I added a few tests as well as a fuzz check and some benchmarks.

Here are the benchmarks results with a naive pair of folds and a `List.equal`:
```
$ dune exec -- bench/bench.exe -quota 5s
Estimated testing time 30s (6 benchmarks x 5s). Change using '-quota'.
┌───────────────────────┬─────────────┬────────────┬──────────┬──────────┬────────────┐
│ Name                  │    Time/Run │    mWd/Run │ mjWd/Run │ Prom/Run │ Percentage │
├───────────────────────┼─────────────┼────────────┼──────────┼──────────┼────────────┤
│ Equal (size 10)       │    290.16ns │    120.00w │          │          │      0.80% │
│ Equal (size 100)      │  2_843.26ns │  1_200.03w │    2.75w │    2.75w │      7.79% │
│ Equal (size 1000)     │ 36_488.74ns │ 12_000.28w │  279.07w │  279.07w │    100.00% │
│ Not equal (size 10)   │    169.39ns │    120.00w │          │          │      0.46% │
│ Not equal (size 100)  │  1_635.65ns │  1_200.03w │    2.74w │    2.74w │      4.48% │
│ Not equal (size 1000) │ 22_713.03ns │ 12_000.28w │  278.92w │  278.92w │     62.25% │
└───────────────────────┴─────────────┴────────────┴──────────┴──────────┴────────────┘
```
and here the results with the actual `equal`:
```
$ dune exec -- bench/bench.exe -quota 5s
Estimated testing time 30s (6 benchmarks x 5s). Change using '-quota'.
┌───────────────────────┬─────────────┬────────────┬──────────┬──────────┬────────────┐
│ Name                  │    Time/Run │    mWd/Run │ mjWd/Run │ Prom/Run │ Percentage │
├───────────────────────┼─────────────┼────────────┼──────────┼──────────┼────────────┤
│ Equal (size 10)       │    245.14ns │    140.01w │          │          │      0.76% │
│ Equal (size 100)      │  2_471.80ns │  1_400.07w │    0.23w │    0.23w │      7.70% │
│ Equal (size 1000)     │ 32_084.88ns │ 14_000.75w │    3.37w │    3.37w │    100.00% │
│ Not equal (size 10)   │     92.12ns │     56.00w │          │          │      0.29% │
│ Not equal (size 100)  │     86.83ns │     98.01w │          │          │      0.27% │
│ Not equal (size 1000) │    113.15ns │    140.01w │          │          │      0.35% │
└───────────────────────┴─────────────┴────────────┴──────────┴──────────┴────────────┘
```

Let me know what you think!