@Julow 2019-04-09
Nice !

This can be a little faster by scarifying the elegant type `enum` and using a list of `node` instead. ([code](https://github.com/NathanReb/ocaml-diet/compare/equality-function...Julow:16_using_list))

```
Before:
┌─────────────────────┬─────────────┬────────────┬──────────┬──────────┬────────────┐
│ Name                │    Time/Run │    mWd/Run │ mjWd/Run │ Prom/Run │ Percentage │
├─────────────────────┼─────────────┼────────────┼──────────┼──────────┼────────────┤
│ Equal/size 10       │    150.96ns │    140.01w │          │          │      0.62% │
│ Equal/size 100      │  1_509.06ns │  1_400.07w │    0.22w │    0.22w │      6.16% │
│ Equal/size 1000     │ 24_513.00ns │ 14_000.75w │    3.45w │    3.45w │    100.00% │
│ Not equal/size 10   │     28.54ns │     42.00w │          │          │      0.12% │
│ Not equal/size 100  │     55.85ns │     98.01w │          │          │      0.23% │
│ Not equal/size 1000 │     77.31ns │    140.01w │          │          │      0.32% │
└─────────────────────┴─────────────┴────────────┴──────────┴──────────┴────────────┘
After:
┌─────────────────────┬─────────────┬───────────┬──────────┬──────────┬────────────┐
│ Name                │    Time/Run │   mWd/Run │ mjWd/Run │ Prom/Run │ Percentage │
├─────────────────────┼─────────────┼───────────┼──────────┼──────────┼────────────┤
│ Equal/size 10       │    135.65ns │    60.00w │          │          │      0.59% │
│ Equal/size 100      │  1_369.21ns │   600.01w │          │          │      5.92% │
│ Equal/size 1000     │ 23_141.91ns │ 6_000.14w │    0.64w │    0.64w │    100.00% │
│ Not equal/size 10   │     24.74ns │    18.00w │          │          │      0.11% │
│ Not equal/size 100  │     48.73ns │    42.00w │          │          │      0.21% │
│ Not equal/size 1000 │     45.17ns │    60.00w │          │          │      0.20% │
└─────────────────────┴─────────────┴───────────┴──────────┴──────────┴────────────┘
```