@NathanReb 2019-04-10
Nice optimization!

I rebased, rewrote the tests a bit differently so it takes advantage of OUnit's nesting and rewrote enum/compare using lists of nodes!